don't use short chunk numbers
Don't be mean with integer sizes.  Although we will probably never
have more than 32k buffered chunks, we should use 32 bit integers for
addressing them.  We do not save very much (some of the saved space is
eaten by alignment anyway), but we save at least one assembler
instruction for converting short to int.

This change requires some more explicit casts, because gcc was less
picky when comparing short with a full int.

git-svn-id: https://svn.musicpd.org/mpd/trunk@7313 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/decode.c
@@ -518,7 +518,7 @@ static void decodeParent(PlayerControl * pc, DecoderControl * dc, OutputBuffer *
 
 		if (do_pause)
 			player_sleep();
-		else if (!outputBufferEmpty(cb) && cb->begin != next) {
+		else if (!outputBufferEmpty(cb) && (int)cb->begin != next) {
 			OutputBufferChunk *beginChunk =
 				outputBufferGetChunk(cb, cb->begin);
 			unsigned int fadePosition;
@@ -564,7 +564,7 @@ static void decodeParent(PlayerControl * pc, DecoderControl * dc, OutputBuffer *
 				break;
 			outputBufferShift(cb);
 			player_wakeup_decoder_nb();
-		} else if (!outputBufferEmpty(cb) && cb->begin == next) {
+		} else if (!outputBufferEmpty(cb) && (int)cb->begin == next) {
 			/* at the beginning of a new song */
 
 			if (doCrossFade == 1 && nextChunk >= 0) {
